Continued updates
Working towards basic functionality for quick demo with Stokke. Added beginning command line call for plink, should be able to automatically connect to the robot. Also had further progress in the actual code writing. Finally, added the basis for the tabs counter aand insertion, for use with logic and loops. Will be adding it to all sections from here on out. Also added a reference scratchpad, for definitions and other information to be kept at hant instead of scrolling through document trying to find a variable name.