Initial functioning code
Got code that will generate output, at least the first few times. Now throwing a index error, but that can be fixed. out.py should be a valid program. Automatic connections are still not working, possibly because I am using the command prompt to launch the program, and it is overriding the program's call of the command prompt to launch plink. Will have to test this with a .jar java executable.