If statements, data types, variable declarations
Major update. Added block definitions and javascript for  strings, numbers, and booleans, as well as an arithmetic operation block and a comparison block.  Finally, an If statement block has been added. These blocks have not been implimented in the LegoCodeGenPython or the blockly.html yet, but they should drop in with few issues. A key decision has been the designation of the | character as the sign to decrement the tab counter when generating python code.