Addition to visable blocks, potential dummyInput issue
Added some of the blocks to the visable portion of the program, and can be added to the work field. However, they have not been added to the LegoCodeGenPython program. Also ran into the issue that our Blockly code does not seem to recognize DummyInput constructors. This is a serious issue, as our logic statements and variable value blocks rely on this. Will be searching for a fix for this in the coming week.