Continued Sensor blocks, start of string split reor
Continued fleshing out the sensors. Occured to me that the old string split system of one run through the loop is not sufficient any more. Need to find a way to use the main splitstring function to split lines into their physical lines, and a secondary split char and loop to spilt and manipulate discrete parts of a line, eg. processing math, sensors, and variables for a single line. No work done on this concept yet, will be a priority in future updates