Repo before secondary loop implimentation
State of the program before implimentation of a second, nested for loop to make it so that we are able to properly parse statements that output a single line of code, but have multiple blocks making up that line. Previous iteration was only meant to handle single blocks per line of code.